refactor: round를 따로 객체화해서 관리
